Run Qt in a dedicated thread
VST plugins must run their event loop in a dedicated thread because
there is no way to integrate with the host's event loop.  The
effEditIdle dispatcher op is only called while the editor GUI is open,
so it would not be possible to run timers or network I/O while the
editor GUI is closed.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>