JamConnection: send full 32-bit usermask instead of just 1 byte
The QByteArray::append(int ch) method was accidentally used to send the
32-bit usermask.

Since explicitly converting to little-endian and casting quint32 to
char* is ugly, create appendLe32() and appendLe16() wrapper functions.
Convert existing methods to use these wrappers to simplify the code and
reduce the chance of similar bugs in the future.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>