RemoteChannel: let RemoteUser connect processAudioStreams()
Follow the same pattern as Metronome and let the owner of RemoteUser
connect processAudioStreams(). RemoteUser actually already calls
connect() so there is no need to do it in RemoteChannel too.

The advantage of letting the owner of RemoteChannel call connect is
looser coupling. For example, a test case could call
processAudioStreams() instead of a real AudioProcessor.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>