Metronome: introduce accent and click audio clips
If an "accent" audio clip is available then the first beat of the
interval plays a different sound than the other beats. This makes it
easier to identify the start of the interval.

A time signature and first beat of a bar audio clip can be added in the
future to make the metronome even more helpful.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>