ChannelStrip: fix null property binding errors on disconnect
The LocalChannel UI control has property bindings to a C++ object. The
Component.destroy() method does not immediately delete the control, so
if the C++ object is deleted in the meantime a null property binding
error occurs. Assign a temporary dummy object to prevent the error and
destroy the control.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>