ChannelStrip: use Repeater instead of ListView
Placing local and remote channels in a ListView never worked well
because the size couldn't be calculated reliably. This was due to how
ListView dynamically instantiates the delegate. I couldn't find a signal
that fires at the appropriate time to update its size.

Luckily the Repeater QML type has the exact behavior needed and the code
is simpler too!

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>