AppView: warn when process tick jitter is too high
The real-time audio engine must be fed data regularly. If the main
thread is unable to regularly produce/consume audio then there will be
underruns/overruns. If this happens it's necessary to identify and fix
the cause of latency and maybe increase the real-time audio buffer size
so more jitter can be tolerated.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>