Metronome: avoid integer truncation sample time bugs
Be careful about converting between float and SampleTime. Integer
truncation can lead to inconsistencies in the sample time calculation.
This caused spurious assertion failures because the first sample of the
interval was sometimes not considered part of the interval.

The assertion failure can be triggered fairly easily with bpi=16 bpm=120
sampleRate=44100 when LocalChannel calls remainingIntervalTime().

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>