build: use qtkeychain CMake dependency support
Probing for the qtkeychain header file works on Linux, but won't work with the
wahjam2-mac-build build environment where the file is not in the system header
directories.

Switch to using CMake to detect the dependency since that's the native
dependency mechanism that qtkeychain supports. Keep the header file probing in
case some systems don't install the CMake support files.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>