qml: fix window sizing on macOS
The sizing is probably not correct yet, because I don't fully understand how to
combine Layouts, anchors, and other positioning/sizing. However, this commit
makes the window resize as expected on macOS. Previously the window sized
itself to fit the login screen and then did not resize itself automatically to
fit the main screen.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>