JamSession: prevent screen sleep while jamming
Do not allow the screen to go into power saving mode while jamming.
While the user is playing their instrument they are not using the mouse
or keyboard that keep the screen from entering its power saving mode.
They need to be able to see what is happening in the jam session, so it
is important that the screen does not turn off.

This code comes from the original Wahjam client and was authored by me.
I have relicensed it under the Apache-2.0 license for Wahjam2.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>