AppView: use processor sample clock for currentSampleTime()
This is not completely correct but it's better than using a
QElapsedTimer as a sample clock. In particular, this doesn't work when
there is a discontinuity in the audio samples (e.g. an
underrun/overrun). The problem with QElapsedTimer is that it's
out-of-sync enough with the AudioProcessor's sample clock that it leads
to inconsistencies in time calculations.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>