ChordChart: show connection status
The ChordChart is a central UI component that has a lot of the user's
attention. It serves no purpose while disconnected from the jam session,
so display the connection status there.

In the future a StackLayout or similar approach could be used to
separate the connection status from the actual ChordChart. It would also
be nice to replace the text descriptions ("Disconnected", etc) with
icons.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>