build: put version number in meson.build for now
The VERSION file was supposed to make it easier to customize the version
in a build. Unfortunately the c++ -I. include directory option causes
any #include <version> to include our VERSION file on macOS where
filenames are case-insensitive.

Get rid of the VERSION file for the time being so that compilation
succeeds on macOS.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>