AudioStream: set ring size to nsamples
Normally each sample does not need its own AudioDescriptor and the ring
size was the number of samples divided by an arbitrary "small buffer
size" factor. This saved memory by allocating fewer AudioDescriptors
that would never be used.

It turns out that it's hard to know what the smallest buffer size will
be. On macOS using a Zoom LiveTrack-8 USB sound card at 44.1 kHz results
in 14-sample buffers and the ring is exhausted because the
SMALL_BLOCK_SIZE (64) factor is insufficient.

Just allocate nsamples AudioDescriptors so that the desired number of
samples will always fit - even if they are written 1 sample at a time.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>